move CategoriesPage components into Category folder
